+++ app/locales/en/translations.ts
@@ -988,8 +988,6 @@ export default {
             project_metadata_tags_label: 'Tags:',
             project_contributors_title: 'Project contributors',
             project_contributors_description: 'Updates made to this section will update the project.',
-            subjects_title: 'Collection subjects',
-            subjects_description: 'Select a discipline and subdiscipline, if relevant. Add more by clicking on a new discipline.',
             collection_metadata_title: 'Collection metadata',
             cancel: 'Cancel',
             add_button: 'Add to collection',